Move back pressure into synchronous call path
