+++ src/batch_write.rs
@@ -28,7 +28,6 @@
 use core::panic;
 use std::sync::Arc;
 
-use log::warn;
 use tokio::runtime::Handle;
 
 use crate::types::RowAttributes;
@@ -52,7 +51,6 @@ pub(crate) struct WriteBatchRequest {
 impl DbInner {
     #[allow(clippy::panic)]
     async fn write_batch(&self, batch: WriteBatch) -> Result<Arc<KVTable>, SlateDBError> {
-        self.maybe_apply_backpressure().await;
         let now = self.options.clock.now();
 
         let current_table = if self.wal_enabled() {
@@ -118,28 +116,6 @@ impl DbInner {
         };
 
         Ok(current_table)
-    }
-
-    pub(crate) async fn maybe_apply_backpressure(&self) {
-        loop {
-            let table = {
-                let guard = self.state.read();
-                let state = guard.state();
-                if state.imm_memtable.len() <= self.options.max_unflushed_memtable {
-                    return;
-                }
-                let Some(table) = state.imm_memtable.back() else {
-                    return;
-                };
-                warn!(
-                    "applying backpressure to write by waiting for imm table flush. imm tables({}), max({})",
-                    state.imm_memtable.len(),
-                    self.options.max_unflushed_memtable
-                );
-                table.clone()
-            };
-            table.await_flush_to_l0().await
-        }
     }
 
     pub(crate) fn spawn_write_task(
